+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file. This projec
     * `Debug`
 * Removed Device Logging Option, which was pushed into new logging under debug.
 * Added Device Logging Override for each Device, by using the Device Config.
+* Change from dependency `public-ip` to `systeminformation`.
 * Housekeeping and updated dependencies.
 
 **Full Changelog**: https://github.com/donavanbecker/homebridge-noip/compare/v1.2.4...v1.3.0